\name{SparseLPM-package}
\alias{SparseLPM-package}
\alias{SparseLPM}
\docType{package}
\title{
\packageTitle{SparseLPM}
}
\description{
\packageDescription{SparseLPM}
}

\author{
Riccardo Rastelli

Mantainer: Riccardo Rastelli <riccardoras@gmail.com>
}
\references{
Rastelli, R. (2018) "The Sparse Latent Position Model for nonnegative weighted networks", \url{https://arxiv.org/abs/1808.09262}
}
\keyword{ package }
